+++ src/templates/tutorial.js
@@ -5,10 +5,10 @@ import { MDXProvider } from "@mdx-js/react"
 import { MDXRenderer } from "gatsby-plugin-mdx"
 import styled from "styled-components"
 import { Twemoji } from "react-emoji-render"
-import Highlight, { defaultProps } from "prism-react-renderer"
 
 import Button from "../components/Button"
 import Card from "../components/Card"
+import Codeblock from "../components/Codeblock"
 import TutorialMetadata from "../components/TutorialMetadata"
 import FileContributors from "../components/FileContributors"
 import InfoBanner from "../components/InfoBanner"
@@ -208,34 +208,7 @@ const components = {
   h3: H3,
   h4: H4,
   h5: H5,
-  // TODO move into component
-  pre: (props) => {
-    const className = props.children.props.className || ""
-    const matches = className.match(/language-(?<lang>.*)/)
-    const language =
-      matches && matches.groups && matches.groups.lang
-        ? matches.groups.lang
-        : ""
-    return (
-      <Highlight
-        {...defaultProps}
-        code={props.children.props.children}
-        language={language}
-      >
-        {({ className, style, tokens, getLineProps, getTokenProps }) => (
-          <pre className={className} style={style}>
-            {tokens.map((line, i) => (
-              <div {...getLineProps({ line, key: i })}>
-                {line.map((token, key) => (
-                  <span {...getTokenProps({ token, key })} />
-                ))}
-              </div>
-            ))}
-          </pre>
-        )}
-      </Highlight>
-    )
-  },
+  pre: Codeblock,
   a: Link,
   Button,
   InfoBanner,
